refactor: remove duplicate snapshot tests, extract shared stress math, add parity tests, tighten error assertions, add accounts.rs coverage
Closes #640
Closes #645
Closes #646
Closes #653
Closes #654
