+++ portfolio-core/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod fx;
 pub mod snapshot;
+pub mod stress;
 pub mod types;
