+++ portfolio-core/src/stress.rs
@@ -1,0 +1,560 @@
+//! Canonical stress-test shock calculation shared by `src-tauri` and `portfolio-mcp`.
+//!
+//! A TypeScript port of this logic lives in `frontend/lib/scenarioMath.ts` for use
+//! when no Tauri backend is available (browser dev mode, scenario-comparison
+//! previews). That copy is hand-maintained and must be kept in sync with this
+//! file if the shock model changes — see #645 / #646.
+
+use crate::types::{PortfolioSnapshot, StressHoldingResult, StressResult, StressScenario};
+
+fn fx_shock_key(currency: &str, base_currency: &str) -> String {
+    format!(
+        "fx_{}_{}",
+        currency.to_lowercase(),
+        base_currency.to_lowercase()
+    )
+}
+
+/// Apply a stress scenario to an existing portfolio snapshot and return the
+/// impact breakdown. This is a pure function — it does not modify the snapshot.
+pub fn run_stress_test(snapshot: &PortfolioSnapshot, scenario: &StressScenario) -> StressResult {
+    let mut holding_results: Vec<StressHoldingResult> = Vec::new();
+    let mut total_stressed = 0.0;
+
+    for holding in &snapshot.holdings {
+        let asset_type_key = holding.asset_type.as_str();
+
+        // Asset-level shock: cash is excluded — only FX shocks may affect cash positions
+        let asset_shock = if asset_type_key == "cash" {
+            0.0
+        } else {
+            scenario.shocks.get(asset_type_key).copied().unwrap_or(0.0)
+        };
+
+        // FX shock: apply if holding is not in the portfolio base currency.
+        let fx_shock = if !holding
+            .currency
+            .eq_ignore_ascii_case(&snapshot.base_currency)
+        {
+            let fx_key = fx_shock_key(&holding.currency, &snapshot.base_currency);
+            scenario.shocks.get(&fx_key).copied().unwrap_or(0.0)
+        } else {
+            0.0
+        };
+
+        let current_value = holding.market_value_cad;
+        // A holding can lose at most its full value — floor at 0 so shocks
+        // beyond -100% (e.g. a -200% scenario) don't drive it negative.
+        let stressed_value = (current_value * (1.0 + asset_shock) * (1.0 + fx_shock)).max(0.0);
+        let impact = stressed_value - current_value;
+
+        // Combined shock for display
+        let shock_applied = (1.0 + asset_shock) * (1.0 + fx_shock) - 1.0;
+
+        holding_results.push(StressHoldingResult {
+            holding_id: holding.id.clone(),
+            symbol: holding.symbol.clone(),
+            name: holding.name.clone(),
+            current_value,
+            stressed_value,
+            impact,
+            shock_applied,
+        });
+
+        total_stressed += stressed_value;
+    }
+
+    let total_stressed = total_stressed.max(0.0);
+    let current_value = snapshot.total_value;
+    let total_impact = total_stressed - current_value;
+    let total_impact_percent = if current_value != 0.0 {
+        (total_impact / current_value) * 100.0
+    } else {
+        0.0
+    };
+
+    StressResult {
+        scenario: scenario.name.clone(),
+        current_value,
+        stressed_value: total_stressed,
+        total_impact,
+        total_impact_percent,
+        holding_breakdown: holding_results,
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::types::{AccountType, AssetType, Holding, HoldingId, HoldingWithPrice};
+    use std::collections::HashMap;
+
+    fn make_holding(
+        symbol: &str,
+        asset_type: AssetType,
+        currency: &str,
+        value: f64,
+    ) -> HoldingWithPrice {
+        HoldingWithPrice {
+            holding: Holding {
+                id: HoldingId(symbol.to_string()),
+                symbol: symbol.to_string(),
+                name: symbol.to_string(),
+                asset_type: asset_type.clone(),
+                account: if matches!(asset_type, AssetType::Cash) {
+                    AccountType::Cash
+                } else {
+                    AccountType::Taxable
+                },
+                account_id: None,
+                account_name: None,
+                quantity: 1.0,
+                cost_basis: value,
+                currency: currency.to_string(),
+                exchange: String::new(),
+                target_weight: None,
+                created_at: "2024-01-01T00:00:00Z".to_string(),
+                updated_at: "2024-01-01T00:00:00Z".to_string(),
+                indicated_annual_dividend: None,
+                indicated_annual_dividend_currency: None,
+                dividend_frequency: None,
+                maturity_date: None,
+            },
+            current_price: value,
+            current_price_cad: value,
+            market_value_cad: value,
+            cost_value_cad: value,
+            gain_loss: 0.0,
+            gain_loss_percent: 0.0,
+            weight: 1.0,
+            target_value: 0.0,
+            target_delta_value: 0.0,
+            target_delta_percent: 0.0,
+            daily_change_percent: 0.0,
+            fx_stale: false,
+            price_is_stale: false,
+        }
+    }
+
+    fn make_snapshot(holdings: Vec<HoldingWithPrice>) -> PortfolioSnapshot {
+        let total = holdings.iter().map(|h| h.market_value_cad).sum();
+        PortfolioSnapshot {
+            holdings,
+            total_value: total,
+            total_cost: total,
+            total_gain_loss: 0.0,
+            total_gain_loss_percent: 0.0,
+            daily_pnl: 0.0,
+            last_updated: "2024-01-01T00:00:00Z".to_string(),
+            base_currency: "CAD".to_string(),
+            total_target_weight: 0.0,
+            target_cash_delta: 0.0,
+            realized_gains: 0.0,
+            annual_dividend_income: 0.0,
+            requires_cost_basis_selection: false,
+        }
+    }
+
+    #[test]
+    fn zero_shocks_return_unchanged_values() {
+        let snapshot = make_snapshot(vec![
+            make_holding("AAPL", AssetType::Stock, "USD", 10_000.0),
+            make_holding("BTC", AssetType::Crypto, "CAD", 5_000.0),
+        ]);
+        let scenario = StressScenario {
+            name: "Zero".to_string(),
+            shocks: HashMap::new(),
+        };
+
+        let result = run_stress_test(&snapshot, &scenario);
+
+        assert!(
+            (result.total_impact).abs() < 0.001,
+            "Zero shocks should produce zero impact"
+        );
+        assert_eq!(result.holding_breakdown.len(), 2);
+        for h in &result.holding_breakdown {
+            assert!((h.impact).abs() < 0.001);
+        }
+    }
+
+    #[test]
+    fn stock_shock_applies_correctly() {
+        let value = 10_000.0;
+        let snapshot = make_snapshot(vec![make_holding("AAPL", AssetType::Stock, "CAD", value)]);
+        let mut shocks = HashMap::new();
+        shocks.insert("stock".to_string(), -0.20);
+        let scenario = StressScenario {
+            name: "Bear".to_string(),
+            shocks,
+        };
+
+        let result = run_stress_test(&snapshot, &scenario);
+
+        let expected_stressed = value * 0.80;
+        assert!((result.stressed_value - expected_stressed).abs() < 0.001);
+        assert!((result.total_impact - (-2_000.0)).abs() < 0.001);
+    }
+
+    #[test]
+    fn fx_shock_applies_to_non_base_holdings() {
+        let value = 10_000.0;
+        let snapshot = make_snapshot(vec![make_holding("AAPL", AssetType::Stock, "USD", value)]);
+        let mut shocks = HashMap::new();
+        shocks.insert("stock".to_string(), -0.10);
+        shocks.insert("fx_usd_cad".to_string(), 0.05);
+        let scenario = StressScenario {
+            name: "Mixed".to_string(),
+            shocks,
+        };
+
+        let result = run_stress_test(&snapshot, &scenario);
+
+        // stressed = 10000 * 0.90 * 1.05 = 9450
+        let expected = value * 0.90 * 1.05;
+        assert!((result.stressed_value - expected).abs() < 0.001);
+    }
+
+    #[test]
+    fn base_currency_holdings_ignore_fx_shock() {
+        let value = 5_000.0;
+        let snapshot = make_snapshot(vec![make_holding("RY.TO", AssetType::Stock, "CAD", value)]);
+        let mut shocks = HashMap::new();
+        shocks.insert("fx_usd_cad".to_string(), 0.15); // should not affect CAD holding
+        let scenario = StressScenario {
+            name: "FX only".to_string(),
+            shocks,
+        };
+
+        let result = run_stress_test(&snapshot, &scenario);
+
+        assert!((result.total_impact).abs() < 0.001);
+    }
+
+    #[test]
+    fn fx_shock_uses_snapshot_base_currency() {
+        let value = 8_000.0;
+        let mut snapshot = make_snapshot(vec![
+            make_holding("RY.TO", AssetType::Stock, "CAD", value),
+            make_holding("MSFT", AssetType::Stock, "USD", value),
+        ]);
+        snapshot.base_currency = "USD".to_string();
+
+        let mut shocks = HashMap::new();
+        shocks.insert("fx_cad_usd".to_string(), -0.10);
+        let scenario = StressScenario {
+            name: "USD Base".to_string(),
+            shocks,
+        };
+
+        let result = run_stress_test(&snapshot, &scenario);
+
+        assert!((result.holding_breakdown[0].stressed_value - 7_200.0).abs() < 0.001);
+        assert!((result.holding_breakdown[1].stressed_value - 8_000.0).abs() < 0.001);
+    }
+
+    #[test]
+    fn cash_ignores_asset_shock_but_applies_fx_shock() {
+        let value = 10_000.0;
+        let snapshot = make_snapshot(vec![
+            make_holding("USD-CASH", AssetType::Cash, "USD", value),
+            make_holding("CAD-CASH", AssetType::Cash, "CAD", value),
+        ]);
+        let mut shocks = HashMap::new();
+        shocks.insert("cash".to_string(), -0.50);
+        shocks.insert("fx_usd_cad".to_string(), 0.10);
+        let scenario = StressScenario {
+            name: "Cash test".to_string(),
+            shocks,
+        };
+
+        let result = run_stress_test(&snapshot, &scenario);
+
+        let usd = result
+            .holding_breakdown
+            .iter()
+            .find(|h| h.symbol == "USD-CASH")
+            .unwrap();
+        let cad = result
+            .holding_breakdown
+            .iter()
+            .find(|h| h.symbol == "CAD-CASH")
+            .unwrap();
+
+        assert!((usd.stressed_value - 11_000.0).abs() < 0.001);
+        assert!((cad.stressed_value - value).abs() < 0.001);
+    }
+
+    #[test]
+    fn empty_snapshot_returns_zero() {
+        let snapshot = make_snapshot(vec![]);
+        let scenario = StressScenario {
+            name: "Empty".to_string(),
+            shocks: HashMap::new(),
+        };
+        let result = run_stress_test(&snapshot, &scenario);
+        assert_eq!(result.total_impact, 0.0);
+        assert_eq!(result.total_impact_percent, 0.0);
+    }
+
+    #[test]
+    fn etf_shock_applies_separately_from_stock_shock() {
+        let snapshot = make_snapshot(vec![
+            make_holding("AAPL", AssetType::Stock, "CAD", 5_000.0),
+            make_holding("VFV", AssetType::Etf, "CAD", 5_000.0),
+        ]);
+        let mut shocks = HashMap::new();
+        shocks.insert("stock".to_string(), -0.10);
+        shocks.insert("etf".to_string(), -0.05);
+        let scenario = StressScenario {
+            name: "Mixed assets".to_string(),
+            shocks,
+        };
+
+        let result = run_stress_test(&snapshot, &scenario);
+
+        let stock = result
+            .holding_breakdown
+            .iter()
+            .find(|h| h.symbol == "AAPL")
+            .unwrap();
+        let etf = result
+            .holding_breakdown
+            .iter()
+            .find(|h| h.symbol == "VFV")
+            .unwrap();
+        assert!((stock.stressed_value - 4_500.0).abs() < 0.001);
+        assert!((etf.stressed_value - 4_750.0).abs() < 0.001);
+    }
+
+    #[test]
+    fn total_loss_scenario_does_not_go_negative() {
+        // A -100% shock wipes out the holding; stressed value should be 0, not negative.
+        let value = 10_000.0;
+        let snapshot = make_snapshot(vec![make_holding("BTC", AssetType::Crypto, "CAD", value)]);
+        let mut shocks = HashMap::new();
+        shocks.insert("crypto".to_string(), -1.0);
+        let scenario = StressScenario {
+            name: "Crypto gone".to_string(),
+            shocks,
+        };
+
+        let result = run_stress_test(&snapshot, &scenario);
+
+        assert!(
+            result.stressed_value >= 0.0,
+            "stressed_value should not be negative"
+        );
+        assert!((result.stressed_value - 0.0).abs() < 0.001);
+        assert!((result.total_impact - (-value)).abs() < 0.001);
+    }
+
+    #[test]
+    fn multi_asset_portfolio_sums_all_shocks_correctly() {
+        let snapshot = make_snapshot(vec![
+            make_holding("AAPL", AssetType::Stock, "CAD", 10_000.0),
+            make_holding("VFV", AssetType::Etf, "CAD", 5_000.0),
+            make_holding("BTC", AssetType::Crypto, "CAD", 2_000.0),
+            make_holding("USD-CASH", AssetType::Cash, "CAD", 3_000.0),
+        ]);
+        let mut shocks = HashMap::new();
+        shocks.insert("stock".to_string(), -0.20);
+        shocks.insert("etf".to_string(), -0.20);
+        shocks.insert("crypto".to_string(), -0.50);
+        let scenario = StressScenario {
+            name: "Bear + crypto winter".to_string(),
+            shocks,
+        };
+
+        let result = run_stress_test(&snapshot, &scenario);
+
+        // stock: 8000, etf: 4000, crypto: 1000, cash: 3000 = 16000
+        let expected_stressed = 8_000.0 + 4_000.0 + 1_000.0 + 3_000.0;
+        assert!((result.stressed_value - expected_stressed).abs() < 0.01);
+        assert_eq!(result.holding_breakdown.len(), 4);
+    }
+
+    #[test]
+    fn scenario_name_is_preserved_in_result() {
+        let snapshot = make_snapshot(vec![make_holding("AAPL", AssetType::Stock, "CAD", 1_000.0)]);
+        let scenario = StressScenario {
+            name: "My Custom Scenario".to_string(),
+            shocks: HashMap::new(),
+        };
+        let result = run_stress_test(&snapshot, &scenario);
+        assert_eq!(result.scenario, "My Custom Scenario");
+    }
+
+    #[test]
+    fn cash_only_portfolio_is_unchanged() {
+        // All holdings are cash — asset shocks must not affect them.
+        let snapshot = make_snapshot(vec![
+            make_holding("CAD-CASH", AssetType::Cash, "CAD", 10_000.0),
+            make_holding("USD-CASH", AssetType::Cash, "CAD", 5_000.0),
+        ]);
+        let mut shocks = HashMap::new();
+        shocks.insert("stock".to_string(), -0.20);
+        shocks.insert("etf".to_string(), -0.10);
+        shocks.insert("crypto".to_string(), -0.50);
+        let scenario = StressScenario {
+            name: "Asset shocks on cash-only portfolio".to_string(),
+            shocks,
+        };
+
+        let result = run_stress_test(&snapshot, &scenario);
+
+        assert!(
+            result.total_impact.abs() < 0.001,
+            "cash holdings must be immune to asset shocks; total_impact={}",
+            result.total_impact
+        );
+        assert!(
+            (result.stressed_value - 15_000.0).abs() < 0.001,
+            "stressed_value should equal original total; got {}",
+            result.stressed_value
+        );
+    }
+
+    #[test]
+    fn zero_cost_basis_holding_no_panic_or_nan() {
+        // A holding with cost_basis = 0.0 must not produce NaN or Inf.
+        let mut holding = make_holding("ZERO", AssetType::Stock, "CAD", 1_000.0);
+        holding.cost_basis = 0.0;
+        // market_value_cad is still non-zero so the stressed value is defined.
+        let snapshot = make_snapshot(vec![holding]);
+        let mut shocks = HashMap::new();
+        shocks.insert("stock".to_string(), -0.10);
+        let scenario = StressScenario {
+            name: "Mild shock on zero cost basis".to_string(),
+            shocks,
+        };
+
+        let result = run_stress_test(&snapshot, &scenario);
+
+        assert!(
+            result.stressed_value.is_finite(),
+            "stressed_value must be finite; got {}",
+            result.stressed_value
+        );
+        assert!(
+            result.total_impact.is_finite(),
+            "total_impact must be finite; got {}",
+            result.total_impact
+        );
+        for h in &result.holding_breakdown {
+            assert!(
+                h.stressed_value.is_finite(),
+                "holding stressed_value must be finite for symbol {}",
+                h.symbol
+            );
+        }
+    }
+
+    #[test]
+    fn extreme_negative_shock_floors_holding_and_portfolio_at_zero() {
+        // Regression guard for #635: a shock beyond -100% (e.g. -200%) must not
+        // drive the stressed value negative. A holding can go to $0 but not below.
+        let value = 10_000.0;
+        let snapshot = make_snapshot(vec![make_holding("BTC", AssetType::Crypto, "CAD", value)]);
+        let mut shocks = HashMap::new();
+        shocks.insert("crypto".to_string(), -2.0); // -200% shock
+        let scenario = StressScenario {
+            name: "Beyond total wipeout".to_string(),
+            shocks,
+        };
+
+        let result = run_stress_test(&snapshot, &scenario);
+
+        assert!(
+            result.stressed_value >= 0.0,
+            "portfolio stressed_value should not be negative, got {}",
+            result.stressed_value
+        );
+        assert!(
+            (result.stressed_value - 0.0).abs() < 0.001,
+            "expected stressed_value floored at 0, got {}",
+            result.stressed_value
+        );
+        assert!(
+            result.holding_breakdown[0].stressed_value >= 0.0,
+            "holding stressed_value should not be negative, got {}",
+            result.holding_breakdown[0].stressed_value
+        );
+        // Total loss is capped at the original value, not 2x the value.
+        assert!((result.total_impact - (-value)).abs() < 0.001);
+    }
+
+    #[test]
+    fn fx_only_shock_no_asset_shocks() {
+        // FX shock on a USD holding raises its CAD value; CAD holding is unaffected.
+        let usd_value = 10_000.0;
+        let cad_value = 5_000.0;
+        let snapshot = make_snapshot(vec![
+            make_holding("AAPL", AssetType::Stock, "USD", usd_value),
+            make_holding("RY.TO", AssetType::Stock, "CAD", cad_value),
+        ]);
+        let mut shocks = HashMap::new();
+        shocks.insert("fx_usd_cad".to_string(), 0.10);
+        let scenario = StressScenario {
+            name: "FX-only: USD strengthens 10%".to_string(),
+            shocks,
+        };
+
+        let result = run_stress_test(&snapshot, &scenario);
+
+        let usd_result = result
+            .holding_breakdown
+            .iter()
+            .find(|h| h.symbol == "AAPL")
+            .unwrap();
+        let cad_result = result
+            .holding_breakdown
+            .iter()
+            .find(|h| h.symbol == "RY.TO")
+            .unwrap();
+
+        // USD holding gains 10%
+        assert!(
+            (usd_result.stressed_value - usd_value * 1.10).abs() < 0.001,
+            "USD holding should gain 10%; expected {} got {}",
+            usd_value * 1.10,
+            usd_result.stressed_value
+        );
+        // CAD holding is unchanged
+        assert!(
+            (cad_result.stressed_value - cad_value).abs() < 0.001,
+            "CAD holding should be unaffected by FX shock; expected {} got {}",
+            cad_value,
+            cad_result.stressed_value
+        );
+    }
+
+    // ── Cross-language parity guard (#646) ──────────────────────────────────
+    // Mirrors the "cross-language parity" test in
+    // frontend/lib/__tests__/scenarioMath.test.ts. Keep the input scenario and
+    // expected numbers identical in both places if either implementation changes.
+    #[test]
+    fn parity_matches_typescript_scenario_math() {
+        let snapshot = make_snapshot(vec![
+            make_holding("AAPL", AssetType::Stock, "USD", 10_000.0),
+            make_holding("RY.TO", AssetType::Stock, "CAD", 5_000.0),
+        ]);
+        let mut shocks = HashMap::new();
+        shocks.insert("stock".to_string(), -0.20);
+        shocks.insert("fx_usd_cad".to_string(), 0.05);
+        let scenario = StressScenario {
+            name: "Parity Check".to_string(),
+            shocks,
+        };
+
+        let result = run_stress_test(&snapshot, &scenario);
+
+        // AAPL: 10000 * 0.80 * 1.05 = 8400
+        assert!((result.holding_breakdown[0].stressed_value - 8_400.0).abs() < 0.000_001);
+        // RY.TO: 5000 * 0.80 = 4000 (CAD holding ignores fx_usd_cad)
+        assert!((result.holding_breakdown[1].stressed_value - 4_000.0).abs() < 0.000_001);
+        assert!((result.stressed_value - 12_400.0).abs() < 0.000_001);
+        assert!((result.total_impact - (-2_600.0)).abs() < 0.000_001);
+        assert!((result.total_impact_percent - (-17.333_333_333_333_332)).abs() < 0.000_001);
+    }
+}
